refactor: merge log into config
