+++ services/interactions/src/commands/config/config.ts
@@ -1,15 +1,16 @@
 import * as interactions from '#interactions';
 import { ConfigCommand } from '#interactions';
-import { ArgumentsOf, send } from '#util';
+import { ArgumentsOf, ControlFlowError, send } from '#util';
 import type {
   ApiGetGuildsSettingsResult,
   ApiPatchGuildSettingsBody,
   ApiPatchGuildSettingsResult,
-  GuildSettings
+  GuildSettings,
+  WebhookToken
 } from '@automoderator/core';
 import { UserPerms } from '@automoderator/discord-permissions';
-import { Rest } from '@automoderator/http-client';
-import { Config, kConfig } from '@automoderator/injection';
+import { Rest, HTTPError as DiscordHTTPError } from '@automoderator/http-client';
+import { Config, kConfig, kSql } from '@automoderator/injection';
 import { Rest as DiscordRest } from '@cordis/rest';
 import { stripIndents } from 'common-tags';
 import {
@@ -18,8 +19,13 @@ import {
   APIGuildInteraction,
   ApplicationCommandPermissionType,
   RESTPutAPIGuildApplicationCommandsPermissionsJSONBody,
-  Routes
+  RESTPostAPIChannelWebhookJSONBody,
+  RESTPostAPIChannelWebhookResult,
+  Routes,
+  ChannelType,
+  Snowflake
 } from 'discord-api-types/v9';
+import type { Sql } from 'postgres';
 import { inject, injectable } from 'tsyringe';
 import { Command } from '../../command';
 import { Handler } from '../../handler';
@@ -32,13 +38,15 @@ export default class implements Command {
     public readonly rest: Rest,
     public readonly discordRest: DiscordRest,
     public readonly handler: Handler,
-    @inject(kConfig) public readonly config: Config
+    @inject(kConfig) public readonly config: Config,
+    @inject(kSql) public readonly sql: Sql<{}>
   ) {}
 
   private async _sendCurrentSettings(interaction: APIGuildInteraction) {
     const settings = await this.rest.get<ApiGetGuildsSettingsResult>(`/guilds/${interaction.guild_id}/settings`);
 
     const atRole = (role?: string | null) => role ? `<@&${role}>` : 'none';
+    const atChannel = (channel?: string | null) => channel ? `<#${channel}>` : 'none';
 
     return send(interaction, {
       content: stripIndents`
@@ -46,6 +54,8 @@ export default class implements Command {
         • mod role: ${atRole(settings.mod_role)}
         • admin role: ${atRole(settings.admin_role)}
         • mute role: ${atRole(settings.mute_role)}
+        • mod logs: ${atChannel(settings.mod_action_log_channel)}
+        • filter logs: ${atChannel(settings.filter_trigger_log_channel)}
         • automatically pardon warnings after: ${settings.auto_pardon_mutes_after ? `${settings.auto_pardon_mutes_after} days` : 'never'}
       `,
       allowed_mentions: { parse: [] }
@@ -57,19 +67,48 @@ export default class implements Command {
       modrole: args.modrole,
       adminrole: args.adminrole,
       muterole: args.muterole,
-      pardon: args.pardonwarnsafter
+      pardon: args.pardonwarnsafter,
+      mod: args.modlogchannel,
+      filters: args.filterslogchannel
     };
   }
 
   public async exec(interaction: APIGuildInteraction, args: ArgumentsOf<typeof ConfigCommand>) {
-    const { modrole, adminrole, muterole, pardon } = this.parse(args);
+    const { modrole, adminrole, muterole, pardon, mod, filters } = this.parse(args);
 
     let settings: Partial<GuildSettings> = {};
 
-    if (modrole) settings.mod_role = modrole.id;
-    if (adminrole) settings.admin_role = adminrole.id;
-    if (muterole) settings.mute_role = muterole.id;
-    if (pardon != null) settings.auto_pardon_mutes_after = pardon;
+    if (modrole) {
+      settings.mod_role = modrole.id;
+    }
+
+    if (adminrole) {
+      settings.admin_role = adminrole.id;
+    }
+
+    if (muterole) {
+      settings.mute_role = muterole.id;
+    }
+
+    if (pardon != null) {
+      settings.auto_pardon_mutes_after = pardon;
+    }
+
+    if (mod) {
+      if (mod.type !== ChannelType.GuildText) {
+        throw new ControlFlowError('Please provide a valid text channel');
+      }
+
+      settings.mod_action_log_channel = mod.id;
+    }
+
+    if (filters) {
+      if (filters.type !== ChannelType.GuildText) {
+        throw new ControlFlowError('Please provide a valid text channel');
+      }
+
+      settings.filter_trigger_log_channel = filters.id;
+    }
 
     if (!Object.values(settings).length) {
       return this._sendCurrentSettings(interaction);
@@ -82,31 +121,6 @@ export default class implements Command {
 
     if (modrole || adminrole) {
       const guild = await this.discordRest.get<APIGuild>(Routes.guild(interaction.guild_id));
-      const permissions: APIApplicationCommandPermission[] = [
-        {
-          id: guild.owner_id,
-          type: ApplicationCommandPermissionType.User,
-          permission: true
-        }
-      ];
-
-      permissions.push(...this.config.devIds.map(id => ({ id, type: ApplicationCommandPermissionType.User, permission: true })));
-
-      if (modrole) {
-        permissions.push({
-          id: modrole.id,
-          type: ApplicationCommandPermissionType.Role,
-          permission: true
-        });
-      }
-
-      if (adminrole) {
-        permissions.push({
-          id: adminrole.id,
-          type: ApplicationCommandPermissionType.Role,
-          permission: true
-        });
-      }
 
       await this.discordRest.put<unknown, RESTPutAPIGuildApplicationCommandsPermissionsJSONBody>(
         Routes.guildApplicationCommandsPermissions(this.config.discordClientId, interaction.guild_id), {
@@ -115,15 +129,122 @@ export default class implements Command {
               ? this.handler.globalCommandIds.get(entry.name)
               : this.handler.testGuildCommandIds.get(`${interaction.guild_id}-${entry.name}`);
 
+            const command = this.handler.commands.get(entry.name);
+
             // eslint-disable-next-line @typescript-eslint/no-unnecessary-condition
-            if ('default_permission' in entry && !entry.default_permission && id) {
-              acc.push({ id, permissions });
+            if ('default_permission' in entry && !entry.default_permission && id && command) {
+              const permissions: APIApplicationCommandPermission[] = [];
+              if (command.userPermissions) {
+                const pushOwner = () => void permissions.push({
+                  id: guild.owner_id,
+                  type: ApplicationCommandPermissionType.User,
+                  permission: true
+                });
+
+                const pushAdmin = () => adminrole && void permissions.push({
+                  id: adminrole.id,
+                  type: ApplicationCommandPermissionType.Role,
+                  permission: true
+                });
+
+                const pushMod = () => modrole && void permissions.push({
+                  id: modrole.id,
+                  type: ApplicationCommandPermissionType.Role,
+                  permission: true
+                });
+
+                switch (command.userPermissions) {
+                  case UserPerms.mod: {
+                    pushMod();
+                    break;
+                  }
+
+                  case UserPerms.admin: {
+                    pushMod();
+                    pushAdmin();
+                    break;
+                  }
+
+                  case UserPerms.owner: {
+                    pushMod();
+                    pushAdmin();
+                    pushOwner();
+                  }
+                }
+              }
+
+              acc.push({
+                id,
+                permissions: [
+                  ...permissions,
+                  ...this.config.devIds.map(id => ({ id, type: ApplicationCommandPermissionType.User, permission: true }))
+                ]
+              });
             }
 
             return acc;
           }, [])
         }
       );
+    }
+
+    const makeWebhook = async (channel: Snowflake, name: string) => {
+      const webhook = await this.discordRest.post<RESTPostAPIChannelWebhookResult, RESTPostAPIChannelWebhookJSONBody>(
+        Routes.channelWebhooks(channel), {
+          data: {
+            name
+          }
+        }
+      ).catch(() => null);
+
+      if (webhook) {
+        const data: WebhookToken = {
+          channel_id: channel,
+          webhook_id: webhook.id,
+          webhook_token: webhook.token!
+        };
+
+        await this.sql`
+          INSERT INTO webhook_tokens ${this.sql(data)}
+          ON CONFLICT (channel_id)
+          DO UPDATE SET ${this.sql(data)}
+        `;
+      }
+    };
+
+    if (settings.mod_action_log_channel) {
+      const [data] = await this.sql<[WebhookToken?]>`SELECT * FROM webhook_tokens WHERE channel_id = ${settings.mod_action_log_channel}`;
+      if (data) {
+        try {
+          await this.discordRest.get(Routes.webhook(data.webhook_id, data.webhook_token));
+        } catch (error) {
+          if (!(error instanceof DiscordHTTPError) || error.response.status !== 404) {
+            throw error;
+          }
+
+          await makeWebhook(settings.mod_action_log_channel, 'Mod Actions');
+        }
+      } else {
+        await makeWebhook(settings.mod_action_log_channel, 'Mod Actions');
+      }
+    }
+
+    if (settings.filter_trigger_log_channel) {
+      const [data] = await this.sql<[WebhookToken?]>`SELECT * FROM webhook_tokens WHERE channel_id = ${settings.filter_trigger_log_channel}`;
+
+      if (data) {
+        try {
+          await this.discordRest.get(Routes.webhook(data.webhook_id, data.webhook_token));
+        } catch (error) {
+          if (!(error instanceof DiscordHTTPError) || error.response.status !== 404) {
+            throw error;
+          }
+
+          await makeWebhook(settings.filter_trigger_log_channel, 'Filter triggers');
+        }
+      } else {
+        await makeWebhook(settings.filter_trigger_log_channel, 'Filter triggers');
+      }
     }
 
     return this._sendCurrentSettings(interaction);
